remove SageMaker vs EC2 difference in batch running
NOTE: We have discovered a weird issue with poppler not running on SageMaker Terminal, this is weird since the Jupyter notebook is able to run poppler. We have tried re-downloading the Anaconda distribution and even re-ran the Anaconda3.sh file to no effect. Hence, we remove the differences between the EC2 and SageMaker instances